Expose Lorekeeper::BacktraceCleaner
