+++ lib/lorekeeper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'lorekeeper/version'
+require 'lorekeeper/backtrace_cleaner'
 require 'lorekeeper/fast_logger'
 require 'lorekeeper/simple_logger'
 require 'lorekeeper/json_logger'
